requirements.txt
pyyaml
colorama
flask
